final product

+++ my_project/backend/migrations/20220927024654_alter.ts
@@ -2,8 +2,8 @@ import { Knex } from "knex";
 
 export async function up(knex: Knex): Promise<void> {
   await knex.schema.alterTable("restaurant", (table) => {
-    table.decimal("lat", 9, 7);
-    table.decimal("lng", 10, 7);
+    table.float("lat", 9, 7);
+    table.float("lng", 10, 7);
   });
 }
